chore: improve type structure
