+++ packages/util/src/index.ts
@@ -3,3 +3,6 @@ export * from './elements.js'
 
 // export everything from the permissions module
 export * from './permissions.js'
+
+// export everything from the types module
+export * from './types.js'
