+++ packages/util/src/permissions.ts
@@ -1,9 +1,8 @@
 import * as DB from '@klicker-uzh/prisma'
-
-// ! do not modify - required for the import of objects not assigned to any catalogue
-// ! NEEDS TO BE CONSISTENT WITH SAME ID IN SHARING SERVICE
-export const MISSING_CATALOG_COLLECTION_ID =
-  'fde06b3c-d515-4907-99cf-c2ba67583155'
+import {
+  MISSING_CATALOG_COLLECTION_ID,
+  type PrismaTransactionClient,
+} from './types.js'
 
 // auxilary type definitions
 type UserAccessMap = {
@@ -13,11 +12,6 @@ type UserAccessMap = {
     derived: boolean
   }
 }
-
-type PrismaTransactionClient = Omit<
-  DB.PrismaClient,
-  '$connect' | '$disconnect' | '$on' | '$transaction' | '$use' | '$extends'
->
 
 // map to directly compare permission levels
 const permissionLevelMap = {
